add functionality so songs are queried and displayed on screen
