+++ client/src/components/App.js
@@ -9,8 +9,8 @@ import Nav from "./navigation/Nav";
 import Register from "./auth/Register";
 import SideBar from './navigation/SideBar';
 import MusicPlayer from "./player/MusicPlayer";
-import Search from './Search';
-import Account from './Account';
+import Search from './navigation/Search';
+import Account from './navigation/Account';
 import SplashPage from "./Splash";
 import CreatePlaylist from "./playlists/new_playlist";
 import PlaylistIndex from "./playlists/playlist_index";
